add new testutilsintegration tag and container helper tests
