+++ tests/Makefile
@@ -1,1 +1,4 @@
 include ../Makefile.Common
+
+GOTEST_OPT+=-count 1
+GO_BUILD_TAGS="testutils"
