+++ tests/testutils/collector_process.go
@@ -172,7 +172,7 @@ func requestConfig(t testing.TB, uri string) map[string]any {
 		if err == nil {
 			break
 		}
-		time.Sleep(1 * time.Second)
+		time.Sleep(2 * time.Second)
 	}
 	require.NoError(t, err)
