+++ tests/testutils/kubeutils/kind_cluster_integration_test.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build testutils
+//go:build testutils && testutilsintegration
 
 package kubeutils
