+++ tests/testutils/testcase.go
@@ -25,7 +25,6 @@ import (
 	"testing"
 	"time"
 
-	docker "github.com/docker/docker/client"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -228,30 +227,6 @@ func (t *Testcase) PrintLogsOnFailure() {
 // Validating shutdown helper for the Testcase's OTLPReceiverSink
 func (t *Testcase) ShutdownOTLPReceiverSink() {
 	require.NoError(t, t.OTLPReceiverSink.Shutdown())
-}
-
-func GetCollectorImageOrSkipTest(t testing.TB) string {
-	image := os.Getenv("SPLUNK_OTEL_COLLECTOR_IMAGE")
-	if strings.TrimSpace(image) == "" {
-		t.Skipf("skipping container-only test (set SPLUNK_OTEL_COLLECTOR_IMAGE env var).")
-		return ""
-	}
-	return image
-}
-
-func CollectorImageIsForArm(t testing.TB) bool {
-	image := os.Getenv("SPLUNK_OTEL_COLLECTOR_IMAGE")
-	if strings.TrimSpace(image) != "" {
-		client, err := docker.NewClientWithOpts(docker.FromEnv)
-		require.NoError(t, err)
-		client.NegotiateAPIVersion(context.Background())
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		inspect, _, err := client.ImageInspectWithRaw(ctx, image)
-		require.NoError(t, err)
-		return inspect.Architecture == "arm64"
-	}
-	return false
 }
 
 // AssertAllLogsReceived is a central helper, designed to avoid most boilerplate. Using the desired
